% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectr.R
\name{create_connection}
\alias{create_connection}
\title{Create a DBI connection to a database.}
\usage{
create_connection(dbname = "main-app")
}
\arguments{
\item{dbname}{character string specifying the database you want to connect to. Use \code{\link{get_databases}} to get a list of available databases.}
}
\description{
Create a DBI connection to a database.
}
