% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectr.R
\name{get_creds}
\alias{get_creds}
\title{Get credentials from AWS}
\usage{
get_creds(dbname = "main-app")
}
\arguments{
\item{dbname}{character string specifying the database you want to get the credentials for. Use \code{\link{get_databases}} to get a list of available databases.}
}
\value{
list with credentials for specified database
}
\description{
Get credentials from AWS
}
