% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectr.R
\name{get_docs}
\alias{get_docs}
\title{Open up the documentation of a database}
\usage{
get_docs(dbname = "main-app", open = TRUE)
}
\arguments{
\item{dbname}{character string specifying the database you want to connect to. Use \code{\link{get_databases}} to get a list of available databases.}

\item{open}{boolean denoting whether or not to open the URL.}
}
\value{
The database documentation URL (invisibly).
}
\description{
Open up the documentation of a database
}
