% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{get_parameter}
\alias{get_parameter}
\title{Get an AWS parameter}
\usage{
get_parameter(name, ...)
}
\arguments{
\item{name}{Name of parameter to retrieve}

\item{...}{Additional arguments to pass to aws. For example, \code{profile = "dbconnect"}.}
}
\description{
Get an AWS parameter
}
