% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{get_parameters}
\alias{get_parameters}
\title{Get AWS parameters}
\usage{
get_parameters(names, ...)
}
\arguments{
\item{names}{Names of parameter to retrieve as a character vector}

\item{...}{Additional arguments to pass to aws. For example, \code{profile = "dbconnect"}.}
}
\description{
Get AWS parameters
}
