% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectr.R
\name{save_creds_yml}
\alias{save_creds_yml}
\title{Save credentials as yaml for a database}
\usage{
save_creds_yml(creds, dbname = "main-app", path = "~/")
}
\arguments{
\item{creds}{\code{\link{get_creds}} to get a list of credentials to save}

\item{dbname}{Use \code{\link{get_databases}} to get a list of available databases to pick one from}

\item{path}{Specify the path to save the credentials to}
}
\description{
Save credentials as yaml for a database
}
