% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnectr.R
\name{update_cred_yamls}
\alias{update_cred_yamls}
\title{Update all local credential files, useful when things need to go fast.}
\usage{
update_cred_yamls(path = "~/")
}
\arguments{
\item{path}{character string specifying the (safe) location for storing the credential files}
}
\description{
Run daily to ensure up-to-date credentials (Old credentials will be valid 48h after rotation).
}
