% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{get_creds}
\alias{get_creds}
\title{Get credentials}
\usage{
get_creds(dbname = "main-app", cache = FALSE,
  cache_folder = "~/.datacamp")
}
\arguments{
\item{dbname}{character string specifying the database you want to get the credentials for. Use \code{\link{get_databases}} to get a list of available databases.}

\item{cache}{boolean that specifies whether or not to fetch and store the credentials in a local cache.}

\item{cache_folder}{if caching is enabled, where to store and fetch the credentials}
}
\value{
list with credentials for specified database
}
\description{
Get credentials
}
